Add clearance csv to inventory items index

+++ app/controllers/inventory_items_controller.rb
@@ -97,15 +97,21 @@ class InventoryItemsController < ApplicationController
   end
 
 
-  def mass_delete
-    @inventory_items = InventoryItem.where(id: params['inventory_item_ids'])
-    @size = @inventory_items.size
-    if @inventory_items.delete_all
-      flash[:notice] = "#{@size} Items Deleted"
+  def mass_action
+    if params['commit'] == 'Delete Items'
+      @inventory_items = InventoryItem.where(id: params['inventory_item_ids'])
+      @size = @inventory_items.size
+      if @inventory_items.delete_all
+        flash[:notice] = "#{@size} Items Deleted"
+      else
+        flash[:error] = 'Items Not Deleted'
+      end
     else
-      flash[:error] = 'Items Not Deleted'
+      inventory_item_ids = params['inventory_item_ids']
+      ClearanceCsvJob.perform_later(User.find(session[:user_id]).email,
+                                 inventory_item_ids)
+      flash[:notice] = "An email with CSV attached will be sent soon"
     end
-
     redirect_to inventory_items_path
   end
